backend/requirements.txt
# backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
pytest==7.4.2